requirements.txt
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
fear-and-greed==0.4
filelock==3.9.0
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.7
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.11.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.2.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
plotly==5.13.1
preprocessing==0.1.13
prometheus-client==0.16.0
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycodestyle==2.10.0
pycoingecko==3.1.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.2
requests-cache==0.9.8
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
scikit-learn==1.2.1
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
sphinx-rtd-theme==0.2.4
stack-data==0.6.1
streamlit==1.22.0
svm==0.1.0
tenacity==8.2.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
tweepy==4.12.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
xmltodict==0.13.0
zipp==3.15.0
